Fixed bugs

+++ otimizacaodespachohidrotermico/OtimizacaoDespachoHidrotermicoGlobals.h
@@ -20,7 +20,7 @@ class OtimizacaoDespachoHidrotermicoGlobals {
 public:
 	vector<int> cascata74;
 	vector<UsinaHidreletrica> hidreletricas;
-	const static double TAXA_DESCONTO = 0.00948879;
+	const static long double TAXA_DESCONTO = 0.00948879;
 	const static int LIMIAR_ERRO_BALANCO_HIDRICO = 16;
 	const static int NUM_PERIODO = 60;
 
@@ -28,11 +28,11 @@ public:
 	static OtimizacaoDespachoHidrotermicoGlobals* obterInstancia();
 	static vector<UsinaTermica> obterUsinasTermicasDoSubsistema(vector<UsinaTermica> usinas, int idUsina);
 	static vector<UsinaHidreletrica> obterUsinasHidreletricasDoSubsistema(vector<UsinaHidreletrica> usinas, int idUsina);
-	static double converterHectometroCubicoParaMetroCubico(double valor, int periodo);
-	static double converterMetroCubicoParaHectometroCubico(double valor, int periodo);
-	static double quantidadeSegundosMes(int periodo);
-	static double quantidadeHorasMes(int periodo);
-	static double quantidadeDiasMes(int periodo);
+	static long double converterHectometroCubicoParaMetroCubico(long double valor, int periodo);
+	static long double converterMetroCubicoParaHectometroCubico(long double valor, int periodo);
+	static long double quantidadeSegundosMes(int periodo);
+	static long double quantidadeHorasMes(int periodo);
+	static long double quantidadeDiasMes(int periodo);
 	static int mesCorrente(int periodo);
 	static UsinaHidreletrica obterUsina(int id_usina);
