+++ usinas/Intercambio.h
@@ -11,15 +11,15 @@
 class Intercambio {
 public:
 	int periodo;
-	double quantidade_subsistema_1;
-	double quantidade_subsistema_2;
-	double quantidade_subsistema_3;
-	double quantidade_subsistema_4;
-	double quantidade_subsistema_5;
+	long double quantidade_subsistema_1;
+	long double quantidade_subsistema_2;
+	long double quantidade_subsistema_3;
+	long double quantidade_subsistema_4;
+	long double quantidade_subsistema_5;
 
 	Intercambio();
-	double totalEnergiaEnviada();
-	double totalEnergiaRecebida(int id_subsistema);
+	long double totalEnergiaEnviada();
+	long double totalEnergiaRecebida(int id_subsistema);
 };
 
 #endif /* INTERCAMBIO_H_ */
