+++ usinas/Usina.h
@@ -20,7 +20,7 @@ public:
 	vector<GeracaoEnergia> geracoes;
 
 	GeracaoEnergia* obterGeracaoEnergia(int periodo);
-	double totalEnergiaGerada();
+	long double totalEnergiaGerada();
 	Usina();
 };
