+++ usinas/UsinaHidreletrica.h
@@ -21,33 +21,33 @@ public:
 	int jusante;
 	vector<int> montantes; //Classe montante Apenas pra poder fazer atributo de UsinaHidreletrica dentro da class UsinaHidreletrica
 
-	double coeficiente_cota_montante_a0;
-	double coeficiente_cota_montante_a1;
-	double coeficiente_cota_montante_a2;
-	double coeficiente_cota_montante_a3;
-	double coeficiente_cota_montante_a4;
+	long double coeficiente_cota_montante_a0;
+	long double coeficiente_cota_montante_a1;
+	long double coeficiente_cota_montante_a2;
+	long double coeficiente_cota_montante_a3;
+	long double coeficiente_cota_montante_a4;
 
-	double coeficiente_cota_jusante_a0;
-	double coeficiente_cota_jusante_a1;
-	double coeficiente_cota_jusante_a2;
-	double coeficiente_cota_jusante_a3;
-	double coeficiente_cota_jusante_a4;
+	long double coeficiente_cota_jusante_a0;
+	long double coeficiente_cota_jusante_a1;
+	long double coeficiente_cota_jusante_a2;
+	long double coeficiente_cota_jusante_a3;
+	long double coeficiente_cota_jusante_a4;
 
-	double tipo_perda_hidraulica;
-	double valor_perda_hidraulica;
+	long double tipo_perda_hidraulica;
+	long double valor_perda_hidraulica;
 
-	double potencia_efetiva;
-	double produtividade_media;
+	long double potencia_efetiva;
+	long double produtividade_media;
 	Reservatorio reservatorio;
 
 	UsinaHidreletrica();
 
 	void atualizarBalancoHidrico(int periodo);
-	double carregar_vazao_montante(int periodo);
-	double carregar_afluencia_montante(int periodo);
-	double calcularGeracaoEnergiaComProdutividadeMedia(int periodo, double volume, double volumeAnterior, double vazaoTurbinada, double vazaoVertida);
-	double calcularPolinomioMontante(double vazaoTotal);
-	double calcularPolinomioJusante(double vazaoTotal);
+	long double carregar_vazao_montante(int periodo);
+	long double carregar_afluencia_montante(int periodo);
+	long double calcularGeracaoEnergiaComProdutividadeMedia(int periodo,long double volume, long double volumeAnterior, long double vazaoTurbinada, long double vazaoVertida);
+	long double calcularPolinomioMontante(long double vazaoTotal);
+	long double calcularPolinomioJusante(long double vazaoTotal);
 };
 
 #endif /* USINAHIDRELETRICA_H_ */
