+++ usinas/UsinaTermica.h
@@ -16,13 +16,13 @@ using namespace std;
 
 class UsinaTermica : public Usina {
 public:
-	double quantidade_geracao_max;
-	double quantidade_geracao_min;
-	double coeficiente_custo_termica_a2;
-	double coeficiente_custo_termica_a1;
-	double coeficiente_custo_termica_a0;
-	double tempo_minimo_ativada;
-	double tempo_minimo_desativada;
+	long double quantidade_geracao_max;
+	long double quantidade_geracao_min;
+	long double coeficiente_custo_termica_a2;
+	long double coeficiente_custo_termica_a1;
+	long double coeficiente_custo_termica_a0;
+	long double tempo_minimo_ativada;
+	long double tempo_minimo_desativada;
 	vector<int> periodos_desativacao_obrigatorio;
 
 	UsinaTermica();
